Move style types of `slider` to `iced_widget`
