+++ style/src/lib.rs
@@ -23,7 +23,6 @@ pub mod pick_list;
 pub mod progress_bar;
 pub mod radio;
 pub mod rule;
-pub mod slider;
 pub mod svg;
 pub mod text_editor;
 pub mod theme;
